server: code deduplication
