+++ cmd/proxy/main.go
@@ -1,9 +1,7 @@
 package main
 
 import (
-	"context"
 	"crypto/rand"
-	"crypto/tls"
 	"encoding/binary"
 	"errors"
 	"flag"
@@ -11,154 +9,39 @@ import (
 	"io"
 	"net"
 	"os"
-	"os/signal"
-	"sync"
-	"syscall"
 	"time"
 
 	"github.com/go-kit/kit/log"
 	quic "github.com/lucas-clemente/quic-go"
 	"github.com/miekg/dns"
-	"github.com/oklog/run"
+
+	server "github.com/ns1/doq-proxy/server"
 )
 
 func main() {
-	l := log.NewLogfmtLogger(log.NewSyncWriter(os.Stdout))
-	l = log.WithPrefix(l, "ts", log.DefaultTimestampUTC)
-
-	var g run.Group
-
-	// proxy code loop
-	{
-		ctx, cancel := context.WithCancel(context.Background())
-		defer cancel()
-
-		g.Add(func() error {
-			return loop(l, ctx)
-		}, func(error) {
-			cancel()
-		})
-	}
-
-	// signal termination
-	{
-		sigterm := make(chan os.Signal, 1)
-		g.Add(func() error {
-			signal.Notify(sigterm, syscall.SIGINT, syscall.SIGTERM)
-			if sig, ok := <-sigterm; ok {
-				l.Log("msg", "stopping the proxy", "signal", sig.String())
-			}
-			return nil
-		}, func(error) {
-			signal.Stop(sigterm)
-			close(sigterm)
-		})
-	}
-
-	err := g.Run()
-	if err != nil {
-		l.Log("msg", "terminating after error", "err", err)
-		os.Exit(1)
-	}
+	server.Main(genParams, handleStream)
 }
 
-func loop(l log.Logger, ctx context.Context) error {
-	var (
-		addr    string
-		tlsCert string
-		tlsKey  string
-		backend string
-	)
+func genParams() server.Params {
+	var params server.Params
 
-	flag.StringVar(&addr, "listen", "127.0.0.1:853", "UDP address to listen on.")
-	flag.StringVar(&tlsCert, "cert", "cert.pem", "TLS certificate path.")
-	flag.StringVar(&tlsKey, "key", "key.pem", "TLS key path.")
+	flag.StringVar(&params.Addr, "listen", "127.0.0.1:853", "UDP address to listen on.")
+	flag.StringVar(&params.TlsCert, "cert", "cert.pem", "TLS certificate path.")
+	flag.StringVar(&params.TlsKey, "key", "key.pem", "TLS key path.")
+
+	var backend string
 	flag.StringVar(&backend, "backend", "8.8.4.4:53", "IP of backend server.")
+	params.Baton = backend
 
 	flag.Parse()
 
-	cert, err := tls.LoadX509KeyPair(tlsCert, tlsKey)
-	if err != nil {
-		return fmt.Errorf("load certificate: %w", err)
-	}
-
-	tls := tls.Config{
-		Certificates: []tls.Certificate{cert},
-		NextProtos:   []string{"doq"},
-	}
-
-	listener, err := quic.ListenAddr(addr, &tls, nil)
-	if err != nil {
-		return fmt.Errorf("listen: %w", err)
-	}
-	defer listener.Close()
-
-	l.Log("msg", "listening for clients", "addr", addr)
-
-	wg := sync.WaitGroup{}
-
-	for {
-		session, err := listener.Accept(ctx)
-		if err != nil {
-			wg.Wait()
-			return fmt.Errorf("accept connection: %w", err)
-		}
-
-		l := log.With(l, "client", session.RemoteAddr())
-		wg.Add(1)
-		go func() {
-			handleClient(l, ctx, session, backend)
-			wg.Done()
-		}()
-	}
-
+	return params
 }
 
-func handleClient(l log.Logger, ctx context.Context, session quic.Connection, backend string) {
-	l.Log("msg", "session accepted")
-
-	var (
-		err error
-		wg  sync.WaitGroup = sync.WaitGroup{}
-	)
-
-	defer func() {
-		msg := ""
-		if err != nil {
-			msg = err.Error()
-		}
-		session.CloseWithError(0, msg)
-
-		l.Log("msg", "session closed")
-	}()
-
-	for {
-		stream, err := session.AcceptStream(ctx)
-		if err != nil {
-			break
-		}
-
-		l := log.With(l, "stream_id", stream.StreamID())
-		l.Log("msg", "stream accepted")
-
-		wg.Add(1)
-		go func() {
-			defer func() {
-				wg.Done()
-				l.Log("msg", "stream closed")
-			}()
-
-			if err := handleStream(stream, backend); err != nil {
-				l.Log("msg", "stream failure", "err", err)
-			}
-		}()
-	}
-
-	wg.Wait()
-}
-
-func handleStream(stream quic.Stream, backend string) error {
+func handleStream(l log.Logger, stream quic.Stream, data any) error {
 	defer stream.Close()
+
+	backend := data.(string)
 
 	wireLength := make([]byte, 2)
 	_, err := io.ReadFull(stream, wireLength)
