+++ go.mod
@@ -1,6 +1,6 @@
 module github.com/ns1/doq-proxy
 
-go 1.17
+go 1.18
 
 require (
 	github.com/go-kit/kit v0.12.0
